Move payment processing to a concern
