+++ core/app/models/spree/order.rb
@@ -10,6 +10,7 @@ module Spree
 
     include Spree::Order::Checkout
     include Spree::Order::CurrencyUpdater
+    include Spree::Order::Payments
 
     checkout_flow do
       go_to_state :address
@@ -416,42 +417,6 @@ module Spree
       @available_payment_methods ||= (PaymentMethod.available(:front_end) + PaymentMethod.available(:both)).uniq
     end
 
-    def pending_payments
-      payments.select { |payment| payment.pending? }
-    end
-
-    def unprocessed_payments
-      payments.select { |payment| payment.checkout? }
-    end
-
-    # processes any pending payments and must return a boolean as it's
-    # return value is used by the checkout state_machine to determine
-    # success or failure of the 'complete' event for the order
-    #
-    # Returns:
-    #
-    # - true if all pending_payments processed successfully
-    #
-    # - true if a payment failed, ie. raised a GatewayError
-    #   which gets rescued and converted to TRUE when
-    #   :allow_checkout_gateway_error is set to true
-    #
-    # - false if a payment failed, ie. raised a GatewayError
-    #   which gets rescued and converted to FALSE when
-    #   :allow_checkout_on_gateway_error is set to false
-    #
-    def process_payments!
-      process_payments_with(:process!)
-    end
-
-    def authorize_payments!
-      process_payments_with(:authorize!)
-    end
-
-    def capture_payments!
-      process_payments_with(:purchase!)
-    end
-
     def billing_firstname
       bill_address.try(:firstname)
     end
@@ -723,21 +688,6 @@ module Spree
         random_token = SecureRandom.urlsafe_base64(nil, false)
         break random_token unless self.class.exists?(guest_token: random_token)
       end
-    end
-
-    def process_payments_with(method)
-      unprocessed_payments.each do |payment|
-        break if payment_total >= total
-
-        payment.public_send(method)
-
-        if payment.completed?
-          self.payment_total += payment.amount
-        end
-      end
-    rescue Core::GatewayError => e
-      result = !!Spree::Config[:allow_checkout_on_gateway_error]
-      errors.add(:base, e.message) and return result
     end
   end
 end
